refactor: added types, updated css, added comments, added proper nameing conventions
