+++ src/components/CodeEditor.css
@@ -1,28 +1,16 @@
 .code-editor {
     display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 50px;
+    height: 100%;
+    width: 100%;
 }
 
 .code-input {
-    width: 100%;
-    height: 200px;
+    width: 50%;
+    height: 100%;
     padding: 10px;
     font-family: 'Courier New', Courier, monospace;
-    font-size: 14px;
+    font-size: 15px;
     border: 1px solid #ccc;
     border-radius: 4px;
     margin-bottom: 20px;
-    resize: none;
-}
-
-.code-output {
-    width: 100%;
-    background: whitesmoke;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    font-family: sans-serif;
-    font-size: 14px;
 }
