--- src/prism.js
@@ -1,5 +1,0 @@
-import Prism from 'prismjs';
-import 'prismjs/components/prism-javascript';
-import 'prismjs/themes/prism.css';
-
-export default Prism;
